BREAKING CHANGE(web-react): Remove old Modal and rename ModalComposed to Modal #DS-617
1. **Discontinue the old `Modal` component:** The old `Modal`
component has been entirely removed and is no longer available
for use.

2. **Adopt the new `Modal` in place of `ModalComposed`:** The
`ModalComposed` component has been renamed to `Modal`. Thus,
replace all instances of `ModalComposed` in your codebase with
`Modal`.

3. **Update Component Tags:** Make sure to modify the component
tags in your codebase as outlined below:

   - `<ModalComposed …>` → `<Modal …>`
   - `<ModalComposedBody …>` → `<ModalBody …>`
   - `<ModalComposedDialog …>` → `<ModalDialog …>`
   - `<ModalComposedFooter …>` → `<ModalFooter …>`
   - `<ModalComposedHeader …>` → `<ModalHeader …>`

Please refer back to this guide or reach out to our team
if you encounter any issues during migration.
