+++ packages/web-react/scripts/entryPoints.js
@@ -20,7 +20,6 @@ const entryPoints = [
   { dirs: ['components', 'Icon'] },
   { dirs: ['components', 'Link'] },
   { dirs: ['components', 'Modal'] },
-  { dirs: ['components', 'ModalComposed'] },
   { dirs: ['components', 'Pagination'] },
   { dirs: ['components', 'Pill'] },
   { dirs: ['components', 'RadioField'] },
