+++ packages/web-react/src/components/Modal/Modal.stories.tsx
@@ -14,6 +14,7 @@ export default {
 } as ComponentMeta<typeof Modal>;
 
 export { default as Modal } from './stories/Modal';
+export { default as ModalWithCustomHeight } from './stories/ModalWithCustomHeight';
 export { default as ModalWithLongText } from './stories/ModalWithLongText';
 export { default as HTML } from './stories/ModalHtml';
 export { default as Props } from './stories/ModalProps';
