+++ packages/web-react/src/components/Modal/ModalFooter.tsx
@@ -1,18 +1,30 @@
 import React from 'react';
 import classNames from 'classnames';
-import { ModalMemberProps } from '../../types';
+import { useStyleProps, useDeprecationMessage } from '../../hooks';
+import { ModalFooterProps } from '../../types';
 import { useModalStyleProps } from './useModalStyleProps';
-import { useStyleProps } from '../../hooks';
 
-const ModalFooter = (props: ModalMemberProps): JSX.Element => {
-  const { children, elementType: ElementTag = 'div', ...restProps } = props;
-  const { classProps } = useModalStyleProps();
+const ModalFooter = (props: ModalFooterProps) => {
+  const { children, alignmentX = 'right', align = 'right', description, ...restProps } = props;
+
+  const { classProps } = useModalStyleProps({ footerAlignment: alignmentX || align });
   const { styleProps, props: otherProps } = useStyleProps(restProps);
 
+  useDeprecationMessage({
+    method: 'property',
+    trigger: !!align,
+    componentName: 'ModalFooter',
+    propertyProps: {
+      deprecatedName: 'align',
+      newName: 'alignmentX',
+    },
+  });
+
   return (
-    <ElementTag {...otherProps} {...styleProps} className={classNames(classProps.footer, styleProps.className)}>
-      {children}
-    </ElementTag>
+    <footer {...otherProps} {...styleProps} className={classNames(classProps.footer.root, styleProps.className)}>
+      {description && <div className={classProps.footer.description}>{description}</div>}
+      <div className={classProps.footer.actions}>{children}</div>
+    </footer>
   );
 };
