+++ packages/web-react/src/components/Modal/__tests__/ModalBody.test.tsx
@@ -4,8 +4,8 @@ import { stylePropsTest } from '../../../../tests/providerTests/stylePropsTest';
 import { restPropsTest } from '../../../../tests/providerTests/restPropsTest';
 import ModalBody from '../ModalBody';
 
-describe('Modal', () => {
-  classNamePrefixProviderTest(ModalBody, 'Modal__body');
+describe('ModalBody', () => {
+  classNamePrefixProviderTest(ModalBody, 'ModalBody');
 
   stylePropsTest(ModalBody);
