+++ packages/web-react/src/components/Modal/__tests__/useModalStyleProps.test.ts
@@ -3,13 +3,15 @@ import { useModalStyleProps } from '../useModalStyleProps';
 
 describe('useModalStyleProps', () => {
   it('should return defaults', () => {
-    const { result } = renderHook(() => useModalStyleProps());
+    const { result } = renderHook(() => useModalStyleProps({}));
 
     expect(result.current.classProps.root).toBe('Modal');
-    expect(result.current.classProps.content).toBe('Modal__content');
-    expect(result.current.classProps.dialog).toBe('Modal__dialog');
-    expect(result.current.classProps.header).toBe('Modal__header');
-    expect(result.current.classProps.body).toBe('Modal__body');
-    expect(result.current.classProps.footer).toBe('Modal__footer');
+    expect(result.current.classProps.dialog).toBe('ModalDialog');
+    expect(result.current.classProps.title).toBe('ModalHeader__title');
+    expect(result.current.classProps.header).toBe('ModalHeader');
+    expect(result.current.classProps.body).toBe('ModalBody');
+    expect(result.current.classProps.footer.root).toBe('ModalFooter ModalFooter--right');
+    expect(result.current.classProps.footer.description).toBe('ModalFooter__description');
+    expect(result.current.classProps.footer.actions).toBe('ModalFooter__actions');
   });
 });
