+++ packages/web-react/src/components/Modal/index.ts
@@ -1,6 +1,7 @@
-export * from './Modal';
 export { default as Modal } from './Modal';
+export { default as ModalDialog } from './ModalDialog';
 export { default as ModalBody } from './ModalBody';
-export { default as ModalFooter } from './ModalFooter';
 export { default as ModalHeader } from './ModalHeader';
+export { default as ModalFooter } from './ModalFooter';
+export * from './ModalContext';
 export * from './useModalStyleProps';
