+++ packages/web-react/src/components/Modal/stories/Modal.tsx
@@ -1,14 +1,18 @@
 // Because there is no `dist` directory during the CI run
 /* eslint-disable import/no-extraneous-dependencies, import/extensions, import/no-unresolved */
-import { ComponentStory } from '@storybook/react';
 import React, { useState } from 'react';
-import { Button } from '../../Button';
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore: No declaration file
+import icons from '@lmc-eu/spirit-icons/dist/icons';
+import { IconsProvider } from '../../../context';
 import Modal from '../Modal';
+import ModalDialog from '../ModalDialog';
 import ModalBody from '../ModalBody';
 import ModalHeader from '../ModalHeader';
-import { Icon } from '../../Icon';
+import ModalFooter from '../ModalFooter';
+import { Button } from '../../Button';
 
-const Story: ComponentStory<typeof Modal> = () => {
+const Story = () => {
   const [isOpen, setOpen] = useState(false);
 
   const toggleModal = () => setOpen(!isOpen);
@@ -18,25 +22,27 @@ const Story: ComponentStory<typeof Modal> = () => {
   };
 
   return (
-    <>
-      <Button onClick={toggleModal} aria-expanded={isOpen} aria-controls="#ModalExample">
-        {isOpen ? 'Close' : 'Open'} Modal
-      </Button>
+    <IconsProvider value={icons}>
+      <div>
+        <Button onClick={toggleModal} aria-expanded={isOpen}>
+          Open Modal
+        </Button>
+      </div>
       <Modal id="ModalExample" isOpen={isOpen} onClose={handleClose}>
-        <ModalHeader>
-          <Button isSquare color="tertiary" onClick={toggleModal} aria-expanded={isOpen} aria-controls="#ModalExample">
-            <Icon name="close" />
-          </Button>
-        </ModalHeader>
-        <ModalBody>
-          <p>
-            Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam at excepturi laudantium magnam mollitia
-            perferendis reprehenderit, voluptate. Cum delectus dicta ducimus eligendi excepturi natus perferendis
-            provident unde. Eveniet, iste, molestiae?
-          </p>
-        </ModalBody>
+        <ModalDialog preferredHeightOnMobile="500px">
+          <ModalHeader>Modal title</ModalHeader>
+          <ModalBody>Body</ModalBody>
+          <ModalFooter>
+            <Button color="primary" onClick={handleClose}>
+              Confirm
+            </Button>
+            <Button color="tertiary" onClick={handleClose}>
+              Cancel
+            </Button>
+          </ModalFooter>
+        </ModalDialog>
       </Modal>
-    </>
+    </IconsProvider>
   );
 };
