--- packages/web-react/src/components/ModalComposed/__tests__/useModalComposedStyleProps.test.ts
@@ -1,17 +1,0 @@
-import { renderHook } from '@testing-library/react-hooks';
-import { useModalComposedStyleProps } from '../useModalComposedStyleProps';
-
-describe('useModalStyleProps', () => {
-  it('should return defaults', () => {
-    const { result } = renderHook(() => useModalComposedStyleProps({}));
-
-    expect(result.current.classProps.root).toBe('Modal Modal--composed');
-    expect(result.current.classProps.dialog).toBe('ModalDialog');
-    expect(result.current.classProps.title).toBe('ModalHeader__title');
-    expect(result.current.classProps.header).toBe('ModalHeader');
-    expect(result.current.classProps.body).toBe('ModalBody');
-    expect(result.current.classProps.footer.root).toBe('ModalFooter ModalFooter--right');
-    expect(result.current.classProps.footer.description).toBe('ModalFooter__description');
-    expect(result.current.classProps.footer.actions).toBe('ModalFooter__actions');
-  });
-});
