--- packages/web-react/src/components/ModalComposed/index.ts
@@ -1,7 +1,0 @@
-export { default as ModalComposed } from './ModalComposed';
-export { default as ModalComposedDialog } from './ModalComposedDialog';
-export { default as ModalComposedBody } from './ModalComposedBody';
-export { default as ModalComposedHeader } from './ModalComposedHeader';
-export { default as ModalComposedFooter } from './ModalComposedFooter';
-export * from './ModalComposedContext';
-export * from './useModalComposedStyleProps';
