--- packages/web-react/src/components/ModalComposed/stories/ModalComposedHtml.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import Html from '../../../../docs/stories/Html';
-
-const Story = () => <Html storyId="modalComposed" groupId="components" />;
-
-export default Story;
