--- packages/web-react/src/components/ModalComposed/stories/ModalComposedProps.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import ModalComposed from '../ModalComposed';
-import Props from '../../../../docs/stories/Props';
-
-const Story = () => <Props component={ModalComposed} />;
-
-export default Story;
