--- packages/web-react/src/components/ModalComposed/stories/ModalComposedWithLongText.tsx
@@ -1,92 +1,0 @@
-// Because there is no `dist` directory during the CI run
-/* eslint-disable import/no-extraneous-dependencies, import/extensions, import/no-unresolved */
-import React, { useState } from 'react';
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore: No declaration file
-import icons from '@lmc-eu/spirit-icons/dist/icons';
-import { IconsProvider } from '../../../context';
-import ModalComposed from '../ModalComposed';
-import ModalComposedDialog from '../ModalComposedDialog';
-import ModalComposedBody from '../ModalComposedBody';
-import ModalComposedHeader from '../ModalComposedHeader';
-import ModalComposedFooter from '../ModalComposedFooter';
-import { Button } from '../../Button';
-
-// @see: https://github.com/storybookjs/storybook/issues/8104#issuecomment-932310244
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const Story = (props: unknown) => {
-  const [isOpen, setOpen] = useState(false);
-
-  const toggleModal = () => setOpen(!isOpen);
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  return (
-    <IconsProvider value={icons}>
-      <div>
-        <Button onClick={toggleModal} aria-expanded={isOpen}>
-          Open Modal
-        </Button>
-      </div>
-      <ModalComposed id="ModalExample" isOpen={isOpen} onClose={handleClose}>
-        <ModalComposedDialog elementType="form" name="jhkjhkj">
-          <ModalComposedHeader>Modal title</ModalComposedHeader>
-          <ModalComposedBody>
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam at excepturi laudantium magnam mollitia
-              perferendis reprehenderit, voluptate. Cum delectus dicta ducimus eligendi excepturi natus perferendis
-              provident unde. Eveniet, iste, molestiae?
-            </p>
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam at excepturi laudantium magnam mollitia
-              perferendis reprehenderit, voluptate. Cum delectus dicta ducimus eligendi excepturi natus perferendis
-              provident unde. Eveniet, iste, molestiae?
-            </p>
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam at excepturi laudantium magnam mollitia
-              perferendis reprehenderit, voluptate. Cum delectus dicta ducimus eligendi excepturi natus perferendis
-              provident unde. Eveniet, iste, molestiae?
-            </p>
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam at excepturi laudantium magnam mollitia
-              perferendis reprehenderit, voluptate. Cum delectus dicta ducimus eligendi excepturi natus perferendis
-              provident unde. Eveniet, iste, molestiae?
-            </p>
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam at excepturi laudantium magnam mollitia
-              perferendis reprehenderit, voluptate. Cum delectus dicta ducimus eligendi excepturi natus perferendis
-              provident unde. Eveniet, iste, molestiae?
-            </p>
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam at excepturi laudantium magnam mollitia
-              perferendis reprehenderit, voluptate. Cum delectus dicta ducimus eligendi excepturi natus perferendis
-              provident unde. Eveniet, iste, molestiae?
-            </p>
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam at excepturi laudantium magnam mollitia
-              perferendis reprehenderit, voluptate. Cum delectus dicta ducimus eligendi excepturi natus perferendis
-              provident unde. Eveniet, iste, molestiae?
-            </p>
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aliquam at excepturi laudantium magnam mollitia
-              perferendis reprehenderit, voluptate. Cum delectus dicta ducimus eligendi excepturi natus perferendis
-              provident unde. Eveniet, iste, molestiae?
-            </p>
-          </ModalComposedBody>
-          <ModalComposedFooter>
-            <Button color="primary" onClick={handleClose}>
-              Confirm
-            </Button>
-            <Button color="tertiary" onClick={handleClose}>
-              Cancel
-            </Button>
-          </ModalComposedFooter>
-        </ModalComposedDialog>
-      </ModalComposed>
-    </IconsProvider>
-  );
-};
-
-export default Story;
