+++ packages/web-react/src/components/index.ts
@@ -15,7 +15,6 @@ export * from './Heading';
 export * from './Icon';
 export * from './Link';
 export * from './Modal';
-export * from './ModalComposed';
 export * from './Pagination';
 export * from './Pill';
 export * from './RadioField';
