+++ packages/web-react/src/types/index.ts
@@ -14,7 +14,6 @@ export * from './label';
 export * from './link';
 export * from './message';
 export * from './modal';
-export * from './modalComposed';
 export * from './pagination';
 export * from './pill';
 export * from './radioField';
